gol-backend/requirements.txt
gunicorn
fastapi
uvicorn
jinja2
SQLAlchemy
PyMySQL
alembic
passlib
python-jose[cryptography]
bcrypt
python-multipart